Update framebuffer and add entity clicking

+++ OpenGLEngine/src/OpenGLEngine/Core/UUID.h
@@ -8,12 +8,12 @@ namespace OpenGLEngine {
     {
     public:
         UUID();
-        UUID(uint64_t uuid);
+        UUID(uint32_t uuid);
         UUID(const UUID&) = default;
 
         static UUID Null() { return UUID(0); }
 
-        operator uint64_t() const { return m_UUID; }
+        operator uint32_t() const { return m_UUID; }
 
         bool UUID::operator==(const UUID& other) const { return m_UUID == other.m_UUID; }
         bool UUID::operator!=(const UUID& other) const { return m_UUID != other.m_UUID; }
@@ -25,7 +25,7 @@ namespace OpenGLEngine {
 			return ss.str();
 		}
     private:
-        uint64_t m_UUID;
+        uint32_t m_UUID;
     };
 
     struct UUIDMapping
@@ -50,7 +50,7 @@ namespace std {
     {
         std::size_t operator()(const OpenGLEngine::UUID& uuid) const
         {
-            return (uint64_t)uuid;
+            return (uint32_t)uuid;
         }
     };
